requirements.txt
streamlit
pandas
pyarrow
requests
plotly
python-dotenv
